+++ packages/sdk/tests/acceptance/transferData.test.ts
@@ -2,15 +2,18 @@ import { describe, expect, it } from 'vitest';
 
 import {
   alphanetRelay,
-  hdx,
-  hydrationAlphanet,
+  laos,
+  laosAlphanet,
+  laosMainnet,
+  maos,
   moonbaseAlpha,
+  moonbeam,
   unit,
 } from '@moonbeam-network/xcm-config';
 import type { AnyParachain, Asset } from '@moonbeam-network/xcm-types';
 import { Sdk, type TransferData } from '../../src';
 import {
-  hydrationAddress,
+  laosMainnetAddress,
   moonEvmAddress,
   substrateAddress,
 } from './constants';
@@ -30,18 +33,25 @@ const transferDateTestConfig: {
     destinationAddress: moonEvmAddress,
   },
   {
-    asset: hdx,
-    source: hydrationAlphanet,
-    sourceAddress: hydrationAddress,
-    destination: moonbaseAlpha,
-    destinationAddress: moonEvmAddress,
-  },
-  {
     asset: unit,
     source: moonbaseAlpha,
     sourceAddress: moonEvmAddress,
     destination: alphanetRelay,
     destinationAddress: substrateAddress,
+  },
+  {
+    asset: laos,
+    source: moonbeam,
+    sourceAddress: moonEvmAddress,
+    destination: laosMainnet,
+    destinationAddress: laosMainnetAddress,
+  },
+  {
+    asset: maos,
+    source: moonbaseAlpha,
+    sourceAddress: moonEvmAddress,
+    destination: laosAlphanet,
+    destinationAddress: laosMainnetAddress,
   },
 ];